Moved update method to header repo Changed datatype fot totalOrderValue to BigDecimal

+++ src/main/java/com/retailhub/ordermanagementservice/model/OrderHeader.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -12,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class OrderHeader {
     private int orderId;
     private int userId;
-    private int totalOrderValue;
+    private BigDecimal totalOrderValue;
     private String orderStatus;
 }
